\name{CDO-package}
\alias{CDO-package}
\alias{CDO}
\docType{package}
\title{
R Codes for CDO Pricing (iTraxx Europe Index Series 8)
}
\description{
This package includes 43 models, introduced in the master thesis CDO, HAME Copulas and an R Package 'CDO', for CDS index tranche pricing under the context of iTraxx Europe Index Series 8.
}
\details{
\tabular{ll}{
Package: \tab CDO\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-07\cr
License: \tab GPL (>= 3) \cr
}
All 43 models can be found in the master thesis CDO, HAME Copulas and an R Package 'CDO' by Yafei Xu at Humboldt-Universit?t zu Berlin.
}
\author{
Yafei Xu

Maintainer: Yafei Xu <yafei.xu@hu-berlin.de>

}
\references{
<https://sites.google.com/site/cdowithr/>
}
\keyword{ CDO }
\keyword{ copulas }
\keyword{ iTraxx Europe }
\keyword{ Credit Risk }
